requirements.txt
Django==5.1.2
django-imagekit==5.0.0
django-mptt==0.16.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
pillow==10.4.0
qrcode==8.0